applications/terra-flow-production/requirements.txt

Flask==3.0.0
gunicorn==21.2.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
numpy==1.26.2
pandas==2.1.3
geopandas==0.14.1
shapely==2.0.2
pytest==7.4.3
python-json-logger==2.0.7
supabase==2.0.3
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
alembic
cryptography
data-conversion
openai
psutil
requests
werkzeug
anthropic
trafilatura
openai
matplotlib
seaborn
reportlab
psutil
